refactor(semantic): consolidate missing-path fallback helper into fs facade
Agent-Logs-Url: https://github.com/SimulatorLife/GMLoop/sessions/094d90f1-874d-4090-ab59-e7a55b731c7b
