+++ src/semantic/src/project-index/fs-facade.ts
@@ -1,5 +1,7 @@
 import { promises as fs } from "node:fs";
 
+import { Core } from "@gmloop/core";
+
 export type ProjectIndexFsFacade = {
     readFile?: (...args: any[]) => Promise<any>;
     writeFile?: (...args: any[]) => Promise<any>;
@@ -19,3 +21,23 @@ export const defaultFsFacade: Required<ProjectIndexFsFacade> = {
     stat: fs.stat,
     readDir: fs.readdir
 };
+
+/**
+ * Run an async filesystem operation and recover only from missing-path errors
+ * (`ENOENT`). All non-missing-path failures are rethrown so callers preserve
+ * existing error semantics while keeping ENOENT fallback handling centralized.
+ */
+export async function runWithMissingPathFallback<TResult>(
+    operation: () => Promise<TResult>,
+    onMissingPath: () => TResult
+): Promise<TResult> {
+    try {
+        return await operation();
+    } catch (error) {
+        if (!Core.isErrorWithCode(error, "ENOENT")) {
+            throw error;
+        }
+
+        return onMissingPath();
+    }
+}
