+++ src/semantic/src/project-index/project-tree.ts
@@ -4,8 +4,7 @@ import path from "node:path";
 import { Core } from "@gmloop/core";
 
 import { createProjectIndexAbortGuard } from "./abort-guard.js";
-import { type ProjectIndexFsFacade } from "./fs-facade.js";
-import { runWithMissingPathFallback } from "./missing-path-fallback.js";
+import { type ProjectIndexFsFacade, runWithMissingPathFallback } from "./fs-facade.js";
 import {
     normalizeProjectFileCategory,
     ProjectFileCategory,
